requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
pydantic-settings
python-dotenv